Remove warnings from tests
