+++ src/maths/matrix_test.rs
@@ -368,27 +368,27 @@ fn a_shearing_transformation_moves_x_in_proportion_to_y() {
 #[test]
 fn individual_transformations_are_applied_in_sequence() {
     let p = Tuple::point(1.0, 0.0, 1.0);
-    let A = Matrix4x4::rotation_x(std::f64::consts::PI / 2.0);
-    let B = Matrix4x4::scaling(5.0, 5.0, 5.0);
-    let C = Matrix4x4::translation(10.0, 5.0, 7.0);
+    let a = Matrix4x4::rotation_x(std::f64::consts::PI / 2.0);
+    let b = Matrix4x4::scaling(5.0, 5.0, 5.0);
+    let c = Matrix4x4::translation(10.0, 5.0, 7.0);
 
-    let p2 = A * p;
+    let p2 = a * p;
     assert_eq!(p2, Tuple::point(1.0, -1.0, 0.0));
-    let p3 = B * p2;
+    let p3 = b * p2;
     assert_eq!(p3, Tuple::point(5.0, -5.0, 0.0));
-    let p4 = C * p3;
+    let p4 = c * p3;
     assert_eq!(p4, Tuple::point(15.0, 0.0, 7.0));
 }
 
 #[test]
 fn chained_transformations_must_be_applied_in_reverse_order() {
     let p = Tuple::point(1.0, 0.0, 1.0);
-    let A = Matrix4x4::rotation_x(std::f64::consts::PI / 2.0);
-    let B = Matrix4x4::scaling(5.0, 5.0, 5.0);
-    let C = Matrix4x4::translation(10.0, 5.0, 7.0);
+    let a = Matrix4x4::rotation_x(std::f64::consts::PI / 2.0);
+    let b = Matrix4x4::scaling(5.0, 5.0, 5.0);
+    let c = Matrix4x4::translation(10.0, 5.0, 7.0);
 
-    let T = C * B * A;
-    assert_eq!(T * p, Tuple::point(15.0, 0.0, 7.0));
+    let t = c * b * a;
+    assert_eq!(t * p, Tuple::point(15.0, 0.0, 7.0));
 }
 
 #[test]
